OTHER: Added 'min_half_window' for optimize_window
Added a 'min_half_window' paremeter for morphological.optimize_window so that small window sizes can be skipped to speed up the calculation. Moved the morphology setup to _algorithm_setup to match the other algorithms. Removed unneeded _smooth function from morphological.
